Cleanup run, enough pain for today.

+++ MODSRC/vazkii/botania/client/core/helper/RenderHelper.java
@@ -115,10 +115,10 @@ public final class RenderHelper {
 		float f1 = 0.00390625F;
 		Tessellator tessellator = Tessellator.instance;
 		tessellator.startDrawingQuads();
-		tessellator.addVertexWithUV((double)(par1 + 0), (double)(par2 + par6), (double)z, (double)((float)(par3 + 0) * f), (double)((float)(par4 + par6) * f1));
-		tessellator.addVertexWithUV((double)(par1 + par5), (double)(par2 + par6), (double)z, (double)((float)(par3 + par5) * f), (double)((float)(par4 + par6) * f1));
-		tessellator.addVertexWithUV((double)(par1 + par5), (double)(par2 + 0), (double)z, (double)((float)(par3 + par5) * f), (double)((float)(par4 + 0) * f1));
-		tessellator.addVertexWithUV((double)(par1 + 0), (double)(par2 + 0), (double)z, (double)((float)(par3 + 0) * f), (double)((float)(par4 + 0) * f1));
+		tessellator.addVertexWithUV(par1 + 0, par2 + par6, z, (par3 + 0) * f, (par4 + par6) * f1);
+		tessellator.addVertexWithUV(par1 + par5, par2 + par6, z, (par3 + par5) * f, (par4 + par6) * f1);
+		tessellator.addVertexWithUV(par1 + par5, par2 + 0, z, (par3 + par5) * f, (par4 + 0) * f1);
+		tessellator.addVertexWithUV(par1 + 0, par2 + 0, z, (par3 + 0) * f, (par4 + 0) * f1);
 		tessellator.draw();
 	}
 }
